Move/rename files and methods
